added expression-based channels to scans. Easy-peasy.  Currently the only names allowed within an expression is the name of raw data columns.  Todo: referencing other channels, reference other scans completely, and referencing slices/joins/multi-dimensional access of higher-dim data.

+++ source/acquaman/AMDacqScanController.cpp
@@ -88,7 +88,7 @@ void AMDacqScanController::onStop()
 		qDebug() << curScan_->detectors().count() << " columns";
 		qDebug() << curScan_->d_.count() << " rows";
 
-		for(int x = 0; x < curScan_->d_.count(); x++){
+		for(int x = 0; unsigned(x) < curScan_->d_.count(); x++){
 			qDebug() << "Starting row " << x;
 			for(int y = 0; y < curScan_->detectors().count(); y++)
 				qDebug() << "Value is " << curScan_->d_.value(y, x);
